Broke solution up into 3 parts: A common core library, a windows GUI and a automated console version that takes only a .config file as input.
