--- Form1.cs
@@ -1,164 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.IO;
-using System.Windows.Forms;
-using System.Diagnostics.Eventing.Reader;
-
-namespace wEventLogSearch
-{
-	public partial class Form1 : Form
-	{
-		public Form1()
-		{
-			InitializeComponent();
-		}
-
-		private void btnSearch_Click(object sender, EventArgs e)
-		{
-			lblErrPath.Text = "";
-			if (string.IsNullOrWhiteSpace(txtBoxEventID.ToString()))
-			{
-				lblResults.Text = "EventID can not be blank";
-				return;
-			}
-			if (!File.Exists(txtBoxEvtFpath.Text) && !Directory.Exists(txtBoxEvtFpath.Text))
-			{
-				lblResults.Text = "Input event log file or folder does not exist: ";
-				lblErrPath.Text = txtBoxEvtFpath.Text;
-				return;
-			}
-			if (!Directory.Exists(Path.GetDirectoryName(txtBoxOutput.Text)))
-			{
-				lblResults.Text = "Output directory does not exist: " + Path.GetDirectoryName(txtBoxOutput.Text);
-				lblErrPath.Text = txtBoxOutput.Text;
-				return;
-			}
-			int i = 0;
-			if (txtValueLocations.Text != "")
-			{
-				string chars = txtValueLocations.Text.Replace(" ", "").Replace(",", "");
-				bool result = int.TryParse(chars, out i);
-				if (result == false)
-				{
-					lblResults.Text = "Value locations must be an integer or comma seperated integer list";
-					lblErrPath.Text = txtValueLocations.Text;
-					return;
-				}
-			}
-			int recordCount = 0;
-			string searchString = "*";
-
-			string directorypath = txtBoxEvtFpath.Text;
-			string writeOutputpath = txtBoxOutput.Text;
-			string filterText = txtboxFilter.Text;
-			bool boolGroupProperties = false;
-
-			if (chkGroupProperties.Checked == true)
-			{
-				boolGroupProperties = true;
-			}
-
-			//set filters for query
-			if (txtBoxEventID.Text != "" && txtBoxEventID.Text != "*")
-			{
-				searchString = "*[System[(EventID=" + txtBoxEventID.Text + ")";
-			}
-			if (txtBoxTimeDiff.Text != "")
-			{
-				long timeFilter = 0;
-				string timeDiff = txtBoxTimeDiff.Text;
-				bool canConvert = long.TryParse(timeDiff, out timeFilter);
-				if (canConvert == true)
-				{
-					if (searchString.Contains("*[System["))
-					{
-						searchString = searchString + " and TimeCreated[timediff(@SystemTime) &lt;= " + timeFilter.ToString() + "]";
-					}
-					else if (searchString.Contains("*") || searchString == "")
-					{
-						searchString = "*[System[TimeCreated[timediff(@SystemTime) <= " + timeFilter.ToString() + "]";
-					}
-				}
-			}
-			if (searchString.Contains("*[System["))
-			{
-				searchString = searchString + "]]";
-			}
-
-			if (radioOptionFile.Checked == true)
-			{
-				//List<EventRecord> foundRecords = EventLogHelper.SearchEventLogs(TxtBoxEvtFpath.Text, searchString);
-				//recordCount = EventLogHelper.WriteEventRecords(foundRecords, writeOutputpath, filterText, true, boolGroupProperties);
-				recordCount = EventLogHelper.SearchEventLog(txtBoxEvtFpath.Text, searchString, writeOutputpath, filterText, boolGroupProperties, txtValueLocations.Text, chkIncludeLogSource.Checked);
-				lblResults.Text = $"{recordCount} results were returned";
-			}
-			else
-			{
-				int recordsCount = 0;
-				string[] fileEntries = Directory.GetFiles(directorypath);
-				foreach (string fileName in fileEntries)
-				{
-					//List<EventRecord> foundRecords = EventLogHelper.SearchEventLogs(fileName, searchString);
-					//recordCount = EventLogHelper.WriteEventRecords(foundRecords, writeOutputpath, filterText, true, boolGroupProperties);
-					recordCount = EventLogHelper.SearchEventLog(fileName, searchString, writeOutputpath, filterText, boolGroupProperties, txtValueLocations.Text, chkIncludeLogSource.Checked);
-					recordsCount = recordsCount + recordCount;
-				}
-				lblResults.Text = $"{recordsCount} results were returned";
-			}
-		}
-
-		string lastInputDirectory = "C:\\";
-		string lastOutputDirectory = "C:\\";
-
-		private void btnInFilePath_Click(object sender, EventArgs e)
-		{
-			if (radioOptionFile.Checked == true)
-			{
-				OpenFileDialog openFileDialog1 = new OpenFileDialog();
-
-				openFileDialog1.InitialDirectory = lastInputDirectory;
-				openFileDialog1.Filter = "evt files (*.evt)|*.evt|evtx files (*.evtx)|*.evtx|All files (*.*)|*.*";
-				openFileDialog1.FilterIndex = 2;
-				openFileDialog1.RestoreDirectory = true;
-
-				if (openFileDialog1.ShowDialog() == DialogResult.OK)
-				{
-					txtBoxEvtFpath.Text = openFileDialog1.FileName;
-					lastInputDirectory = Path.GetDirectoryName(openFileDialog1.FileName);
-				}
-			}
-			else
-			{
-				FolderBrowserDialog folderBrowserDialog1 = new System.Windows.Forms.FolderBrowserDialog();
-				folderBrowserDialog1.SelectedPath = lastInputDirectory;
-				if (folderBrowserDialog1.ShowDialog() == DialogResult.OK)
-				{
-					txtBoxEvtFpath.Text = folderBrowserDialog1.SelectedPath;
-					lastInputDirectory = Path.GetDirectoryName(folderBrowserDialog1.SelectedPath);
-				}
-			}
-		}
-
-		private void btnOutFilePath_Click(object sender, EventArgs e)
-		{
-			SaveFileDialog SaveFileDialog1 = new SaveFileDialog();
-
-			SaveFileDialog1.InitialDirectory = lastOutputDirectory;
-			SaveFileDialog1.Filter = "text files (*.txt)|*.txt|csv files (*.csv)|*.csv|All files (*.*)|*.*";
-			SaveFileDialog1.FilterIndex = 2;
-			SaveFileDialog1.RestoreDirectory = true;
-
-			if (SaveFileDialog1.ShowDialog() == DialogResult.OK)
-			{
-				txtBoxOutput.Text = SaveFileDialog1.FileName;
-				lastOutputDirectory = Path.GetDirectoryName(SaveFileDialog1.FileName);
-			}
-		}
-	}
-}
