+++ WELSConsole/App.config
@@ -1,0 +1,21 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+	<startup>
+		<supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8" />
+	</startup>
+	<appSettings>
+
+		<add key="Path.Input" value="C:\Temp\wEventLogSearch\InputEventFiles.evtx"/>
+		<add key="Path.Output" value="C:\Temp\wEventLogSearch\EventLogSearch.csv"/>
+
+		<add key="Search.EventIDs" value="4689 or EventID=4688"/>
+		<add key="Search.Filter" value=""/>
+		<add key="Search.ValueLocations" value=""/>
+		<add key="Search.TimeDifference" value="-1"/>
+
+		<add key="Bool.InputIsSingleFile" value="True"/>
+		<add key="Bool.IncludeLogSource" value="True"/>
+		<add key="Bool.GroupIntoOneColumn" value="False"/>
+
+	</appSettings>
+</configuration>
